grades

+++ WebApplication8/Controllers/ProgramCoursesController.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
@@ -13,62 +14,103 @@ namespace WebApplication8.Controllers
     public class ProgramCoursesController : Controller
     {
         private readonly ApplicationDbContext _context;
+        private readonly UserManager<IdentityUser> _userManager;
 
-        public ProgramCoursesController(ApplicationDbContext context)
+        public ProgramCoursesController(ApplicationDbContext context, UserManager<IdentityUser> userManager)
         {
             _context = context;
+            _userManager = userManager;
+        }
+
+        // Helper method to get the logged-in user's managed study program
+        private async Task<int?> GetManagedStudyProgramId()
+        {
+            var userId = _userManager.GetUserId(User);
+            if (User.IsInRole("Admin"))
+                return null; // Admin can access everything
+
+            var manager = await _context.ProgramManager
+                .FirstOrDefaultAsync(pm => pm.UserId == userId);
+
+            return manager?.StudyProgramId;
         }
 
         // GET: ProgramCourses
         public async Task<IActionResult> Index()
         {
-            var applicationDbContext = _context.ProgramCourses.Include(p => p.Course).Include(p => p.CourseType).Include(p => p.StudyProgram);
-            return View(await applicationDbContext.ToListAsync());
+            var managedProgramId = await GetManagedStudyProgramId();
+
+            var query = _context.ProgramCourses
+                .Include(p => p.Course)
+                .Include(p => p.CourseType)
+                .Include(p => p.StudyProgram)
+                .AsQueryable();
+
+            if (managedProgramId.HasValue)
+            {
+                query = query.Where(pc => pc.StudyProgramId == managedProgramId.Value);
+            }
+
+            return View(await query.ToListAsync());
         }
 
         // GET: ProgramCourses/Details/5
         public async Task<IActionResult> Details(int? id)
         {
-            if (id == null)
-            {
-                return NotFound();
-            }
+            if (id == null) return NotFound();
 
             var programCourse = await _context.ProgramCourses
                 .Include(p => p.Course)
                 .Include(p => p.CourseType)
                 .Include(p => p.StudyProgram)
                 .FirstOrDefaultAsync(m => m.StudyProgramId == id);
-            if (programCourse == null)
-            {
-                return NotFound();
-            }
+
+            if (programCourse == null) return NotFound();
+
+            var managedProgramId = await GetManagedStudyProgramId();
+            if (managedProgramId.HasValue && programCourse.StudyProgramId != managedProgramId.Value)
+                return Forbid();
 
             return View(programCourse);
         }
 
         // GET: ProgramCourses/Create
-        public IActionResult Create()
+        public async Task<IActionResult> Create()
         {
+            var managedProgramId = await GetManagedStudyProgramId();
+
+            if (managedProgramId.HasValue)
+            {
+                ViewData["StudyProgramId"] = new SelectList(
+                    _context.StudyPrograms.Where(sp => sp.StudyProgramId == managedProgramId.Value),
+                    "StudyProgramId", "Name");
+            }
+            else
+            {
+                ViewData["StudyProgramId"] = new SelectList(_context.StudyPrograms, "StudyProgramId", "Name");
+            }
+
             ViewData["CourseId"] = new SelectList(_context.Courses, "CourseId", "CourseId");
             ViewData["CourseTypeId"] = new SelectList(_context.CourseTypes, "CourseTypeId", "Name");
-            ViewData["StudyProgramId"] = new SelectList(_context.StudyPrograms, "StudyProgramId", "Name");
             return View();
         }
 
         // POST: ProgramCourses/Create
-        // To protect from overposting attacks, enable the specific properties you want to bind to.
-        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Create([Bind("StudyProgramId,CourseId,CourseTypeId")] ProgramCourse programCourse)
         {
+            var managedProgramId = await GetManagedStudyProgramId();
+            if (managedProgramId.HasValue && programCourse.StudyProgramId != managedProgramId.Value)
+                return Forbid();
+
             if (ModelState.IsValid)
             {
                 _context.Add(programCourse);
                 await _context.SaveChangesAsync();
                 return RedirectToAction(nameof(Index));
             }
+
             ViewData["CourseId"] = new SelectList(_context.Courses, "CourseId", "CourseId", programCourse.CourseId);
             ViewData["CourseTypeId"] = new SelectList(_context.CourseTypes, "CourseTypeId", "Name", programCourse.CourseTypeId);
             ViewData["StudyProgramId"] = new SelectList(_context.StudyPrograms, "StudyProgramId", "Name", programCourse.StudyProgramId);
@@ -78,16 +120,15 @@ namespace WebApplication8.Controllers
         // GET: ProgramCourses/Edit/5
         public async Task<IActionResult> Edit(int? id)
         {
-            if (id == null)
-            {
-                return NotFound();
-            }
+            if (id == null) return NotFound();
 
             var programCourse = await _context.ProgramCourses.FindAsync(id);
-            if (programCourse == null)
-            {
-                return NotFound();
-            }
+            if (programCourse == null) return NotFound();
+
+            var managedProgramId = await GetManagedStudyProgramId();
+            if (managedProgramId.HasValue && programCourse.StudyProgramId != managedProgramId.Value)
+                return Forbid();
+
             ViewData["CourseId"] = new SelectList(_context.Courses, "CourseId", "CourseId", programCourse.CourseId);
             ViewData["CourseTypeId"] = new SelectList(_context.CourseTypes, "CourseTypeId", "Name", programCourse.CourseTypeId);
             ViewData["StudyProgramId"] = new SelectList(_context.StudyPrograms, "StudyProgramId", "Name", programCourse.StudyProgramId);
@@ -95,16 +136,15 @@ namespace WebApplication8.Controllers
         }
 
         // POST: ProgramCourses/Edit/5
-        // To protect from overposting attacks, enable the specific properties you want to bind to.
-        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Edit(int id, [Bind("StudyProgramId,CourseId,CourseTypeId")] ProgramCourse programCourse)
         {
-            if (id != programCourse.StudyProgramId)
-            {
-                return NotFound();
-            }
+            if (id != programCourse.StudyProgramId) return NotFound();
+
+            var managedProgramId = await GetManagedStudyProgramId();
+            if (managedProgramId.HasValue && programCourse.StudyProgramId != managedProgramId.Value)
+                return Forbid();
 
             if (ModelState.IsValid)
             {
@@ -115,17 +155,12 @@ namespace WebApplication8.Controllers
                 }
                 catch (DbUpdateConcurrencyException)
                 {
-                    if (!ProgramCourseExists(programCourse.StudyProgramId))
-                    {
-                        return NotFound();
-                    }
-                    else
-                    {
-                        throw;
-                    }
+                    if (!ProgramCourseExists(programCourse.StudyProgramId)) return NotFound();
+                    else throw;
                 }
                 return RedirectToAction(nameof(Index));
             }
+
             ViewData["CourseId"] = new SelectList(_context.Courses, "CourseId", "CourseId", programCourse.CourseId);
             ViewData["CourseTypeId"] = new SelectList(_context.CourseTypes, "CourseTypeId", "Name", programCourse.CourseTypeId);
             ViewData["StudyProgramId"] = new SelectList(_context.StudyPrograms, "StudyProgramId", "Name", programCourse.StudyProgramId);
@@ -135,20 +170,19 @@ namespace WebApplication8.Controllers
         // GET: ProgramCourses/Delete/5
         public async Task<IActionResult> Delete(int? id)
         {
-            if (id == null)
-            {
-                return NotFound();
-            }
+            if (id == null) return NotFound();
 
             var programCourse = await _context.ProgramCourses
                 .Include(p => p.Course)
                 .Include(p => p.CourseType)
                 .Include(p => p.StudyProgram)
                 .FirstOrDefaultAsync(m => m.StudyProgramId == id);
-            if (programCourse == null)
-            {
-                return NotFound();
-            }
+
+            if (programCourse == null) return NotFound();
+
+            var managedProgramId = await GetManagedStudyProgramId();
+            if (managedProgramId.HasValue && programCourse.StudyProgramId != managedProgramId.Value)
+                return Forbid();
 
             return View(programCourse);
         }
@@ -159,12 +193,17 @@ namespace WebApplication8.Controllers
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
             var programCourse = await _context.ProgramCourses.FindAsync(id);
+
+            var managedProgramId = await GetManagedStudyProgramId();
+            if (managedProgramId.HasValue && programCourse.StudyProgramId != managedProgramId.Value)
+                return Forbid();
+
             if (programCourse != null)
             {
                 _context.ProgramCourses.Remove(programCourse);
+                await _context.SaveChangesAsync();
             }
 
-            await _context.SaveChangesAsync();
             return RedirectToAction(nameof(Index));
         }
